+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Update docca submodule
 * Add version of create() with uid parameter
+* Add make_salt and make_uid to quickref
 
 ---